+++ src/private/package/GetNovaPackageMetadataElement.ps1
@@ -1,0 +1,15 @@
+function Get-NovaPackageMetadataElement {
+    [CmdletBinding()]
+    param(
+        [Parameter(Mandatory)][string]$Name,
+        [AllowNull()][string]$Value
+    )
+
+    if ( [string]::IsNullOrWhiteSpace($Value)) {
+        return $null
+    }
+
+    $escapedValue = [System.Security.SecurityElement]::Escape($Value)
+    return "    <$Name>$escapedValue</$Name>"
+}
+